[purge.c] fixed warning (hope it is ok now)
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1770 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/purge.c
@@ -651,10 +651,11 @@ renumber (int *nprimes, hashtable_t *H, char *sos)
             {
               if (GET_HASH_R(H,i) == minus2)
                 fprintf (stderr, "Warning: singleton rational prime %llu\n",
-                         GET_HASH_P(H,i));
+                         (unsigned long long) GET_HASH_P(H,i));
               else
                 fprintf (stderr, "Warning: singleton algebraic ideal (%llu,%llu)\n",
-                         GET_HASH_P(H,i), GET_HASH_R(H,i));
+                         (unsigned long long) GET_HASH_P(H,i),
+                         (unsigned long long) GET_HASH_R(H,i));
             }
           H->hashcount[i] = nb++;
           if (fsos != NULL)
